+++ packages/router/test/lib/helpers/prepare.spec.ts
@@ -1,4 +1,6 @@
 import { expect, jsonifyError } from "@connext/nxtp-utils";
+import { BigNumber } from "ethers";
+import { parseEther, parseUnits } from "ethers/lib/utils";
 import { AmountInvalid } from "../../../src/lib/errors/prepare";
 import {
   getReceiverAmount,
@@ -32,6 +34,21 @@ describe("getReceiverAmount", () => {
   it("should work", async () => {
     const result = await getReceiverAmount("10000", 1, 1);
     expect(result).to.be.eq((10000 * 0.9995).toString());
+  });
+
+  it("should work for 6 to 18", async () => {
+    const result = await getReceiverAmount(parseUnits("1", 6).toString(), 6, 18);
+    expect(result).to.be.eq(parseEther("0.9995").toString());
+  });
+
+  it("should work for 18 to 6", async () => {
+    const result = await getReceiverAmount(parseUnits("1", 18).toString(), 18, 6);
+    expect(result).to.be.eq(parseUnits("0.9995", 6).toString());
+  });
+
+  it("should work for decimals", async () => {
+    const result = await getReceiverAmount("104731245", 6, 6);
+    expect(() => BigNumber.from(result)).to.not.throw();
   });
 
   it("should fail if its a decimal string", async () => {