+++ packages/sdk/src/subgraph/subgraph.ts
@@ -1,4 +1,4 @@
-import { providers, Signer } from "ethers";
+import { providers } from "ethers";
 import {
   createLoggingContext,
   FallbackSubgraph,
@@ -113,7 +113,7 @@ export class Subgraph {
   private chainConfig: Record<number, SubgraphChainConfig>;
 
   constructor(
-    private readonly user: Signer,
+    private readonly userAddress: Promise<string>,
     _chainConfig: Record<number, Omit<SubgraphChainConfig, "subgraphSyncBuffer"> & { subgraphSyncBuffer?: number }>,
     private readonly logger: Logger,
     skipPolling = false,
@@ -334,7 +334,7 @@ export class Subgraph {
     const txs = await Promise.all(
       Object.keys(this.sdks).map(async (c) => {
         try {
-          const user = (await this.user.getAddress()).toLowerCase();
+          const user = (await this.userAddress).toLowerCase();
           const chainId = parseInt(c);
           const subgraph = this.sdks[chainId];
 
@@ -544,7 +544,7 @@ export class Subgraph {
 
     const fulfilledTxs = await Promise.all(
       Object.keys(this.sdks).map(async (c) => {
-        const user = (await this.user.getAddress()).toLowerCase();
+        const user = (await this.userAddress).toLowerCase();
         const chainId = parseInt(c);
         const subgraph = this.sdks[chainId];
 
@@ -638,7 +638,7 @@ export class Subgraph {
 
     const cancelledTxs = await Promise.all(
       Object.keys(this.sdks).map(async (c) => {
-        const user = (await this.user.getAddress()).toLowerCase();
+        const user = (await this.userAddress).toLowerCase();
         const chainId = parseInt(c);
         const subgraph = this.sdks[chainId];