+++ packages/sdk/src/transactionManager/transactionManager.ts
@@ -1,23 +1,22 @@
-import { BigNumber, constants, Contract, providers, Signer } from "ethers";
+import { BigNumber, constants, providers, Contract } from "ethers";
 import {
   PrepareParams,
   CancelParams,
   FulfillParams,
-  isNode,
   Logger,
   RequestContext,
   createLoggingContext,
+  GAS_ESTIMATES,
 } from "@connext/nxtp-utils";
-import { parseError } from "@connext/nxtp-txservice";
 import { TransactionManager as TTransactionManager, IERC20Minimal } from "@connext/nxtp-contracts/typechain";
+import { parseError } from "@connext/nxtp-txservice";
 import TransactionManagerArtifact from "@connext/nxtp-contracts/artifacts/contracts/TransactionManager.sol/TransactionManager.json";
 import ERC20 from "@connext/nxtp-contracts/artifacts/contracts/interfaces/IERC20Minimal.sol/IERC20Minimal.json";
+import { Interface } from "ethers/lib/utils";
 import contractDeployments from "@connext/nxtp-contracts/deployments.json";
 
 import { ChainNotConfigured, PriceOracleNotConfigured } from "../error";
 import { getDecimals, getTokenPrice } from "../utils";
-
-const HARDCODED_GAS_LIMIT = BigNumber.from(125_000);
 
 /**
  * Returns the address of the `TransactionManager` deployed to the provided chain, or undefined if it has not been deployed
@@ -78,13 +77,15 @@ export class TransactionManager {
   private chainConfig: {
     [chainId: number]: {
       provider: providers.FallbackProvider;
-      transactionManager: TTransactionManager;
+      transactionManagerAddress: string;
       priceOracleAddress: string;
     };
   };
 
+  private txManagerInterface = new Interface(TransactionManagerArtifact.abi) as TTransactionManager["interface"];
+  private erc20Interface = new Interface(ERC20.abi) as IERC20Minimal["interface"];
+
   constructor(
-    private readonly signer: Signer,
     _chainConfig: {
       [chainId: number]: {
         provider: providers.FallbackProvider;
@@ -92,17 +93,13 @@ export class TransactionManager {
         priceOracleAddress: string;
       };
     },
+    private readonly signerAddress: Promise<string>,
     private readonly logger: Logger,
   ) {
     this.chainConfig = {};
     Object.entries(_chainConfig).forEach(([chainId, { provider, transactionManagerAddress, priceOracleAddress }]) => {
-      const transactionManager = new Contract(
-        transactionManagerAddress,
-        TransactionManagerArtifact.abi,
-        provider,
-      ) as TTransactionManager;
       this.chainConfig[parseInt(chainId)] = {
-        transactionManager,
+        transactionManagerAddress,
         provider,
         priceOracleAddress,
       };
@@ -115,21 +112,8 @@ export class TransactionManager {
    * @param chainId - The chain you want the address on
    * @returns The deployed address or `undefined` if it has not been deployed yet
    */
-  getTransactionManagerAddress(chainId: number): string {
-    return this.chainConfig[chainId].transactionManager.address;
-  }
-
-  /**
-   * Returns a connected signer. This is necessary because the browser-injected signer will not allow a provider to be connected.
-   *
-   * @param provider - The provider to connect if not a browser
-   * @returns The connected signer
-   */
-  getConnectedSigner(provider: providers.FallbackProvider): Signer {
-    if (isNode()) {
-      return this.signer.connect(provider);
-    }
-    return this.signer;
+  getTransactionManagerAddress(chainId: number): string | undefined {
+    return this.chainConfig[chainId]?.transactionManagerAddress;
   }
 
   /**
@@ -149,7 +133,7 @@ export class TransactionManager {
     chainId: number,
     prepareParams: PrepareParams,
     _requestContext?: RequestContext<string>,
-  ): Promise<providers.TransactionResponse> {
+  ): Promise<providers.TransactionRequest> {
     const { requestContext, methodContext } = createLoggingContext(
       "TransactionManager.prepare",
       _requestContext,
@@ -158,16 +142,14 @@ export class TransactionManager {
 
     this.logger.info("Method start", requestContext, methodContext, { chainId, prepareParams });
 
-    const { transactionManager, provider } = this.chainConfig[chainId] ?? {};
-    if (!transactionManager || !provider) {
+    const { provider, transactionManagerAddress } = this.chainConfig[chainId] ?? {};
+    if (!provider || !transactionManagerAddress) {
       throw new ChainNotConfigured(chainId, Object.keys(this.chainConfig));
     }
 
     const { txData, amount, expiry, encodedBid, bidSignature, encryptedCallData } = prepareParams;
 
-    const signer = this.getConnectedSigner(provider);
-
-    const invariant = {
+    const invariantData = {
       receivingChainTxManagerAddress: txData.receivingChainTxManagerAddress,
       user: txData.user,
       router: txData.router,
@@ -183,40 +165,26 @@ export class TransactionManager {
       transactionId: txData.transactionId,
     };
 
-    const connected = transactionManager.connect(signer);
+    const data = this.txManagerInterface.encodeFunctionData("prepare", [
+      {
+        invariantData,
+        amount,
+        expiry,
+        encryptedCallData,
+        encodedBid,
+        bidSignature,
+        encodedMeta: "0x",
+      },
+    ]);
 
-    const contractArgs = {
-      invariantData: invariant,
-      amount,
-      expiry,
-      encryptedCallData,
-      encodedBid,
-      bidSignature,
-      encodedMeta: "0x",
-    };
+    this.logger.info("Prepare transaction created", requestContext, methodContext);
 
-    // estimate gas
-    let gasLimit;
-    try {
-      if (chainId === 100) {
-        gasLimit = HARDCODED_GAS_LIMIT;
-      } else {
-        gasLimit = await connected.estimateGas.prepare(contractArgs);
-      }
-    } catch (e) {
-      const sanitized = parseError(e);
-      throw sanitized;
-    }
-
-    const tx = await connected.prepare(contractArgs, {
+    return {
+      to: transactionManagerAddress,
       value: txData.sendingAssetId === constants.AddressZero ? BigNumber.from(amount) : constants.Zero,
-      from: this.signer.getAddress(),
-      gasLimit: gasLimit.mul(2),
-    });
-    this.logger.info("Prepare transaction submitted", requestContext, methodContext, {
-      txHash: tx.hash,
-    });
-    return tx;
+      data,
+      from: await this.signerAddress,
+    };
   }
 
   /**
@@ -236,7 +204,7 @@ export class TransactionManager {
     chainId: number,
     cancelParams: CancelParams,
     _requestContext?: RequestContext<string>,
-  ): Promise<providers.TransactionResponse> {
+  ): Promise<providers.TransactionRequest> {
     const { requestContext, methodContext } = createLoggingContext(
       "TransactionManager.cancel",
       _requestContext,
@@ -245,45 +213,30 @@ export class TransactionManager {
 
     this.logger.info("Method start", requestContext, methodContext, { cancelParams });
 
-    const { transactionManager, provider } = this.chainConfig[chainId] ?? {};
-    if (!transactionManager || !provider) {
+    const { transactionManagerAddress, provider } = this.chainConfig[chainId] ?? {};
+    if (!transactionManagerAddress || !provider) {
       throw new ChainNotConfigured(chainId, Object.keys(this.chainConfig));
     }
 
     const { txData, signature } = cancelParams;
-    const signer = this.getConnectedSigner(provider);
 
-    const connected = transactionManager.connect(signer);
+    this.logger.info("Cancel transaction created", requestContext, methodContext);
 
-    const contractArgs = {
-      txData,
-      signature,
-      encodedMeta: "0x",
+    const data = this.txManagerInterface.encodeFunctionData("cancel", [
+      {
+        txData,
+        signature,
+        encodedMeta: "0x",
+      },
+    ]);
+
+    this.logger.info("Prepare transaction created", requestContext, methodContext);
+
+    return {
+      to: transactionManagerAddress,
+      data,
+      from: await this.signerAddress,
     };
-
-    // estimate gas
-    let gasLimit;
-    try {
-      if (chainId === 100) {
-        gasLimit = HARDCODED_GAS_LIMIT;
-      } else {
-        gasLimit = await connected.estimateGas.cancel(contractArgs, {
-          from: this.signer.getAddress(),
-        });
-      }
-    } catch (e) {
-      const sanitized = parseError(e);
-      throw sanitized;
-    }
-
-    const tx = await transactionManager
-      .connect(signer)
-      .cancel(contractArgs, { from: this.signer.getAddress(), gasLimit: gasLimit.mul(2) });
-
-    this.logger.info("Cancel transaction submitted", requestContext, methodContext, {
-      txHash: tx.hash,
-    });
-    return tx;
   }
 
   /**
@@ -305,7 +258,7 @@ export class TransactionManager {
     chainId: number,
     fulfillParams: FulfillParams,
     _requestContext?: RequestContext<string>,
-  ): Promise<providers.TransactionResponse> {
+  ): Promise<providers.TransactionRequest> {
     const { requestContext, methodContext } = createLoggingContext(
       "TransactionManager.fulfill",
       _requestContext,
@@ -314,46 +267,30 @@ export class TransactionManager {
 
     this.logger.info("Method start", requestContext, methodContext, { fulfillParams });
 
-    const { transactionManager, provider } = this.chainConfig[chainId] ?? {};
-    if (!transactionManager || !provider) {
+    const { transactionManagerAddress, provider } = this.chainConfig[chainId] ?? {};
+    if (!transactionManagerAddress || !provider) {
       throw new ChainNotConfigured(chainId, Object.keys(this.chainConfig));
     }
 
     const { txData, relayerFee, signature, callData } = fulfillParams;
-    const signer = this.getConnectedSigner(provider);
 
-    const connected = transactionManager.connect(signer);
+    const data = this.txManagerInterface.encodeFunctionData("fulfill", [
+      {
+        txData,
+        relayerFee,
+        signature,
+        callData,
+        encodedMeta: "0x",
+      },
+    ]);
 
-    const contractArgs = {
-      txData,
-      relayerFee,
-      signature,
-      callData,
-      encodedMeta: "0x",
+    this.logger.info("Fulfill transaction created", requestContext, methodContext);
+
+    return {
+      to: transactionManagerAddress,
+      data,
+      from: await this.signerAddress,
     };
-
-    // estimate gas
-    let gasLimit;
-    try {
-      if (chainId === 100) {
-        gasLimit = HARDCODED_GAS_LIMIT;
-      } else {
-        gasLimit = await connected.estimateGas.fulfill(contractArgs, {
-          from: this.signer.getAddress(),
-        });
-      }
-    } catch (e) {
-      const sanitized = parseError(e);
-      throw sanitized;
-    }
-
-    const tx = await connected.fulfill(contractArgs, {
-      from: this.signer.getAddress(),
-      gasLimit: gasLimit.mul(2),
-    });
-
-    this.logger.info("Fulfill transaction submitted", requestContext, methodContext, { txHash: tx.hash });
-    return tx;
   }
 
   /**
@@ -372,7 +309,7 @@ export class TransactionManager {
     amount: string,
     infiniteApprove = false,
     _requestContext?: RequestContext,
-  ): Promise<providers.TransactionResponse | undefined> {
+  ): Promise<providers.TransactionRequest | undefined> {
     const { requestContext, methodContext } = createLoggingContext(
       "TransactionManager.approveTokensIfNeeded",
       _requestContext,
@@ -380,43 +317,32 @@ export class TransactionManager {
 
     this.logger.info("Method start", requestContext, methodContext, { chainId, assetId, amount });
 
-    const { transactionManager, provider } = this.chainConfig[chainId] ?? {};
-    if (!transactionManager || !provider) {
+    const { transactionManagerAddress, provider } = this.chainConfig[chainId] ?? {};
+    if (!transactionManagerAddress || !provider) {
       throw new ChainNotConfigured(chainId, Object.keys(this.chainConfig));
     }
 
-    const signerAddress = await this.signer.getAddress();
-    const signer = this.getConnectedSigner(provider);
-    const erc20 = new Contract(assetId, ERC20.abi, signer) as IERC20Minimal;
-
-    const approved = await erc20.allowance(signerAddress, transactionManager.address);
+    const approvedData = this.erc20Interface.encodeFunctionData("allowance", [
+      await this.signerAddress,
+      transactionManagerAddress,
+    ]);
+    const approvedEncoded = await provider.call({
+      to: assetId,
+      data: approvedData,
+    });
+    const [approved] = this.erc20Interface.decodeFunctionResult("allowance", approvedEncoded);
     this.logger.info("Got approved tokens", requestContext, methodContext, { approved: approved.toString() });
-    if (approved.lt(amount)) {
-      // estimate gas
-      let gasLimit;
-      try {
-        if (chainId === 100) {
-          gasLimit = HARDCODED_GAS_LIMIT;
-        } else {
-          gasLimit = await erc20.estimateGas.approve(
-            transactionManager.address,
-            infiniteApprove ? constants.MaxUint256 : amount,
-            {
-              from: this.signer.getAddress(),
-            },
-          );
-        }
-      } catch (e) {
-        const sanitized = parseError(e);
-        throw sanitized;
-      }
-
-      const tx = await erc20.approve(transactionManager.address, infiniteApprove ? constants.MaxUint256 : amount, {
-        from: this.signer.getAddress(),
-        gasLimit: gasLimit.mul(2),
-      });
-      this.logger.info("Approve transaction submitted", requestContext, methodContext, { txHash: tx.hash });
-      return tx;
+    if (BigNumber.from(approved).lt(amount)) {
+      const data = this.erc20Interface.encodeFunctionData("approve", [
+        transactionManagerAddress,
+        infiniteApprove ? constants.MaxUint256 : amount,
+      ]);
+      this.logger.info("Approve transaction created", requestContext, methodContext);
+      return {
+        to: assetId,
+        data,
+        from: await this.signerAddress,
+      };
     } else {
       this.logger.info("Allowance sufficient", requestContext, methodContext, {
         approved: approved.toString(),
@@ -435,12 +361,16 @@ export class TransactionManager {
    * @returns Either the BigNumber representation of the available router liquidity in the provided asset, or a TransactionManagerError if the function failed
    */
   async getRouterLiquidity(chainId: number, router: string, assetId: string): Promise<BigNumber> {
-    const { transactionManager } = this.chainConfig[chainId] ?? {};
-    if (!transactionManager) {
+    const { transactionManagerAddress, provider } = this.chainConfig[chainId] ?? {};
+    if (!transactionManagerAddress || !provider) {
       throw new ChainNotConfigured(chainId, Object.keys(this.chainConfig));
     }
 
-    return transactionManager.routerBalances(router, assetId);
+    const data = this.txManagerInterface.encodeFunctionData("routerBalances", [router, assetId]);
+    const encoded = await provider.call({ to: transactionManagerAddress, data });
+    const [balance] = this.txManagerInterface.decodeFunctionResult("routerBalances", encoded);
+
+    return BigNumber.from(balance);
   }
 
   /**
@@ -460,13 +390,25 @@ export class TransactionManager {
       fulfillParams.txData.transactionId,
     );
 
-    const { transactionManager, provider } = this.chainConfig[chainId] ?? {};
-    if (!transactionManager || !provider) {
+    const { transactionManagerAddress, provider } = this.chainConfig[chainId] ?? {};
+    if (!transactionManagerAddress || !provider) {
       throw new ChainNotConfigured(chainId, Object.keys(this.chainConfig));
     }
 
     this.logger.info("Method start", requestContext, methodContext, { fulfillParams });
-    const gasAmount = await this.calculateGasAmountForFulfill(chainId, fulfillParams);
+    const gasAmount = BigNumber.from(GAS_ESTIMATES.fulfill); // hardcode gas estimates for now
+    // TODO: this does not account for calldata. we need to figure out how to estimate that later
+    // issue is that we cannot estimate it in the auction request because the tx is not prepared yet
+    // however we might be able to work around this by directly estimating the callto with the calldata
+
+    let gasPrice = BigNumber.from(0);
+    try {
+      gasPrice = await provider.getGasPrice();
+    } catch (e) {
+      const sanitized = parseError(e);
+      throw sanitized;
+    }
+
     const { txData } = fulfillParams;
 
     const ethPriceInUsd = await getTokenPrice(
@@ -492,6 +434,7 @@ export class TransactionManager {
     const outputDecimals = await getDecimals(txData.receivingAssetId, provider);
 
     const tokenAmount = gasAmount
+      .mul(gasPrice)
       .mul(ethPriceInUsd)
       .div(receivingTokenPriceInUsd)
       .div(BigNumber.from(10).pow(18 - outputDecimals));
@@ -506,13 +449,17 @@ export class TransactionManager {
    * @param fulfillParams The params used for fulfill transactions
    */
   async calculateGasAmountForFulfill(chainId: number, fulfillParams: FulfillParams): Promise<BigNumber> {
-    const { transactionManager, provider } = this.chainConfig[chainId] ?? {};
-    if (!transactionManager || !provider) {
+    const { transactionManagerAddress, provider } = this.chainConfig[chainId] ?? {};
+    if (!transactionManagerAddress || !provider) {
       throw new ChainNotConfigured(chainId, Object.keys(this.chainConfig));
     }
 
-    const signer = this.getConnectedSigner(provider);
-    const connected = transactionManager.connect(signer);
+    const transactionManager = new Contract(
+      transactionManagerAddress,
+      TransactionManagerArtifact.abi,
+      provider,
+    ) as TTransactionManager;
+
     const { txData, relayerFee, signature, callData } = fulfillParams;
 
     const contractArgs = {
@@ -526,7 +473,7 @@ export class TransactionManager {
     // get gas limit
     let gasLimit = BigNumber.from(0);
     try {
-      gasLimit = await connected.estimateGas.fulfill(contractArgs);
+      gasLimit = await transactionManager.estimateGas.fulfill(contractArgs);
     } catch (e) {
       const sanitized = parseError(e);
       throw sanitized;