+++ packages/sdk/src/utils.ts
@@ -3,6 +3,7 @@ import {
   generateMessagingInbox as _generateMessagingInbox,
   recoverAuctionBid as _recoverAuctionBid,
   signFulfillTransactionPayload as _signFulfillTransactionPayload,
+  getFulfillTransactionHashToSign as _getFulfillTransactionHashToSign,
   ERC20Abi,
   getOnchainBalance as _getOnchainBalance,
   getNtpTimeSeconds,
@@ -63,17 +64,17 @@ export const getDecimals = async (assetId: string, provider: providers.FallbackP
 /**
  * Gets token price in usd.
  *
- * @param oralceAddress The price oracle address
+ * @param oracleAddress The price oracle address
  * @param tokenAddress The token address to get the price
  *
  * @returns price in usd by decimals 18.
  */
 export const getTokenPrice = async (
-  oralceAddress: string,
+  oracleAddress: string,
   tokenAddress: string,
   provider: providers.FallbackProvider,
 ): Promise<BigNumber> => {
-  const priceOracleContract = new Contract(oralceAddress, PriceOracleAbi, provider);
+  const priceOracleContract = new Contract(oracleAddress, PriceOracleAbi, provider);
   const tokenPriceInBigNum = await priceOracleContract.getTokenPrice(tokenAddress);
   return tokenPriceInBigNum;
 };
@@ -84,6 +85,8 @@ export const getTokenPrice = async (
  *
  */
 export const signFulfillTransactionPayload = _signFulfillTransactionPayload;
+
+export const getFulfillTransactionHashToSign = _getFulfillTransactionHashToSign;
 
 /**
  * This is only here to make it easier for sinon mocks to happen in the tests. Otherwise, this is a very dumb thing.